+++ src/scene/Scene.cc
@@ -295,7 +295,7 @@ void Scene::run() {
 			if (_update) {
 
 				auto updateStartTime = time();
-				(_update)(*this, (float)runT, (float)deltaRunT);
+				(_update)(*this, runT, deltaRunT);
 				UpdateUserTimeStats(_stats, updateStartTime, time());
 			}
