+++ tests/030-rigidbody/main.cc
@@ -40,10 +40,10 @@ constexpr float					PHYSICS_TIMESTEP =		1.0/240.0;
 constexpr bool					DARK =					true;
 
 
-void UpdateCallback(Scene& scene, float time, float deltaTime);
-void WillRenderCallback(VisualWorld& world, float time, float deltaTime);
-void DidRenderCallback(VisualWorld& world, float time, float deltaTime);
-void DidSimulatePhysicsCallback(PhysicalWorld& world, float time, float deltaTime);
+void UpdateCallback(Scene& scene, double time, double deltaTime);
+void WillRenderCallback(VisualWorld& world, double time, double deltaTime);
+void DidRenderCallback(VisualWorld& world, double time, double deltaTime);
+void DidSimulatePhysicsCallback(PhysicalWorld& world, double time, double deltaTime);
 
 
 void InitLog();
@@ -355,7 +355,7 @@ int main(int argc, const char* argv[]) {
 	Scene Callbacks
  ***************************************************************************************/
 
-void UpdateCallback(Scene& scene, float time, float deltaTime) {
+void UpdateCallback(Scene& scene, double time, double deltaTime) {
 	LOG_T(g_logger, "scene: {:p}, time: {}, deltaTime: {}", (void*)&scene, time, deltaTime);
 
 	auto inputManager = scene.inputManager();
@@ -676,20 +676,20 @@ void UpdateCallback(Scene& scene, float time, float deltaTime) {
 				}
 
 				if (keysDown.count(Key::W) || mouseButtonsDown.count(MouseButton::Four)) {
-					vec3 positionDelta = deltaTime * MOVE_SPEED * moveMultiplier * camForward;
+					vec3 positionDelta = (float)deltaTime * MOVE_SPEED * moveMultiplier * camForward;
 					pov->position(pov->position() + positionDelta);
 				}
 				else if (keysDown.count(Key::S)) {
-					vec3 positionDelta = deltaTime * MOVE_SPEED * moveMultiplier * -camForward;
+					vec3 positionDelta = (float)deltaTime * MOVE_SPEED * moveMultiplier * -camForward;
 					pov->position(pov->position() + positionDelta);
 				}
 
 				if (keysDown.count(Key::A)) {
-					vec3 positionDelta = deltaTime * MOVE_SPEED * moveMultiplier * -camRight;
+					vec3 positionDelta = (float)deltaTime * MOVE_SPEED * moveMultiplier * -camRight;
 					pov->position(pov->position() + positionDelta);
 				}
 				else if (keysDown.count(Key::D)) {
-					vec3 positionDelta = deltaTime * MOVE_SPEED * moveMultiplier * camRight;
+					vec3 positionDelta = (float)deltaTime * MOVE_SPEED * moveMultiplier * camRight;
 					pov->position(pov->position() + positionDelta);
 				}
 
@@ -698,7 +698,7 @@ void UpdateCallback(Scene& scene, float time, float deltaTime) {
 					if (keysDown.count(Key::LeftShift)) {
 						direction = -1;
 					}
-					vec3 positionDelta = deltaTime * MOVE_SPEED * moveMultiplier * camUp;
+					vec3 positionDelta = (float)deltaTime * MOVE_SPEED * moveMultiplier * camUp;
 					pov->position(pov->position() + positionDelta * direction);
 				}
 			}
@@ -710,11 +710,11 @@ void UpdateCallback(Scene& scene, float time, float deltaTime) {
 	VisualWorld Callbacks
  ***************************************************************************************/
 
-void WillRenderCallback(VisualWorld& world, float time, float deltaTime) {
+void WillRenderCallback(VisualWorld& world, double time, double deltaTime) {
 	LOG_T(g_logger, "world: {:p}, time: {}, deltaTime: {}", (void*)&world, time, deltaTime);
 }
 
-void DidRenderCallback(VisualWorld& world, float time, float deltaTime) {
+void DidRenderCallback(VisualWorld& world, double time, double deltaTime) {
 	LOG_T(g_logger, "world: {:p}, time: {}, deltaTime: {}", (void*)&world, time, deltaTime);
 }
 
@@ -722,7 +722,7 @@ void DidRenderCallback(VisualWorld& world, float time, float deltaTime) {
 	PhysicalWorld Callbacks
  ***************************************************************************************/
 
-void DidSimulatePhysicsCallback(PhysicalWorld& world, float time, float deltaTime) {
+void DidSimulatePhysicsCallback(PhysicalWorld& world, double time, double deltaTime) {
 	LOG_T(g_logger, "world: {:p}, time: {}, deltaTime: {}", (void*)&world, time, deltaTime);
 }
